Walmart Plots
Assignment: Suppose you are the chief data analyst for Walmart. The CEO has tasked you to select a new town for expansion. You are provided with data from different towns which already have a walmart location. This data contains the population and profits (loss if value is negative) of each town. 
a) Plot the given data using matplotlib, profit on y-axis and population on x-axis. Include your graph in your report.
b) Use linear regression to fit a line to the data which will be used for prediction. You can use available machine learning libraries (sklearn, numpy etc. )
c) Plot the generated line using matplotlib and include the plot in your report.
d) What can you say about the nature of the line?
e) To prepare for a board meeting at Walmart, the CEO has given you a list of cities and their population. Use your machine learning algorithm to rank the cities in descending order of profit potential. In your report, included the selected cities and their estimated revenue in the correct order.
f) What is the accuracy of your model?