Facebook Network Graph
Assignment: You will build a graph that represents the relationship between Facebook users. There is an edge between two users if they are friends on Facebook. The weight of the edge is the closeness of two users, which is defined as the number of communications in each common post. Using this graph, you will be able to connect two unacquainted users with those friends’ connections. You can also estimate the potential probability of them becoming new friends through their close friends.